setupConnection: use bmc.Vendor() instead of bmc.Model()
Instead of having a per hardware model default password,
we set one for each vendor in bmcbutler.yml.

Since .Vendor() is a constant its not likely to fail,
and we can attempt logins with the vendor default credentials.

This change requires bmcbutler.yml to be updated to declare
vendor default passwords instead of model default.