discover: optional hint for probing
Add an functional option that parametrizes and retrieves per callback an opaque hint for the most recent successful probe to use.  That hint is a simple `int` that indexes into the static slice of probe functions.  The probe indices are implementation detail and thus the `int` is opaque.  If the a persisted hint gets out of sync with the implementation or if the machine behind the coordinate that is probed changes to a different model, the hinting either just makes the probing sub-optimal (hint to a non-matching probe) or reverts to default behaviour (hint out of bounds).

The functional option pattern is used in order to keep the default behaviour (no hinting) simple and to keep the published function signature of `discover.ScanAndCoonect()` backwards compatible.