discover: optional hint for probing
Add a functional options that parametrize and/or retrieve per callback a
hint for the most recent successful probe to use.  That hint is a
predefined string constant that points to a specific connect probe.  If
the persisted hint gets out of sync with the implementation or if the
machine behind the coordinate that is probed changes to a different
model, the hinting just makes the probing sub-optimal (hint to a
non-matching probe) or reverts to default behaviour (hint does not match
any probe).

The testing is cleaned up and uses a test case table with parallelized
subtests.  Reflection became necessary, because type assertions are not
parametrizable.  Every mock BMC is tested with no hint, empty hint and
all available hints.