Move provider registration error check:
To fix the nil pointer error that can happen if an error
is not checked from a providers InitFn, we skip adding the
provider to the registry. This should happen in setProviders
not in getProvider. This way we limit the times we call
initFn. Plus setProviders already puts the provider interface
into the registry. getProviders does not need to call initFn
again, it can just get the interface from the client.Registry.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>