Update ipmi provider IsCompatible Func:
Checking if a machine is on for the compatibility
check will cause the check to fail if the BMC is up,
is compatible, but the host machine is powered off.
Just checking is a power state call returned successful
should allow the check to be more accurate and work even
if the host machine is off.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>