Fix bug in provider ipmitool:
Power off in ipmitool was duplicating the internal ipmi
package functionality in PowerOff but doing it poorly.
Specifically not checking for an error when calling `IsOn`.

Also, the internal ipmi package `PowerOn` has been updated.
It now follow how ipmitool behaves when calling `chassis power on`.
ipmitool will return an exit code of 0 if calling `chassis power on`
against a machine in a powered on state.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>