Update metadata user experience:
After using the metadata for a little, i dont like the
experience. Having to pass in metadata was not easily
understood and was trying to be too clever. This update
will make metadata available for all methods. user just
has to call GetMetadata().

Public API is affected as client will no longer be able
to pass in the &metadata var. Seeing as we are not v1,
i feel comfortable with the change.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>