client: additional parameter for connection timeouts
The context deadline timeout value is shared between all the bmclib providers
that is - when creating an a new connection/session on a BMC. The shared timeout value
entails that a single provider can consume the whole timeout value depending on how long
it takes to respond, thus leaving the other providers with no chance to connect.

 WithConnectTimeout sets the timeout when connecting to a BMC to create a new session.
When not defined the deadline value passed in through the context applies.

This timeout value will best work when its lesser than the passed in context deadline value.