Add per provider timeouts for all methods:
This expands the timeout for opening connections
to all method/BMC interactions.

If a context with a timeout IS provided to a method call
but no per provider timeout is defined, then we default to a
timeout of the provided timeout duration divided by the
number of providers/drivers in the Client.Registry.Drivers
slice.

If a context with a timeout IS NOT provided to a method call
and no per provider timeout is defined, then we use the
defaultConnectTimeout of 30 seconds.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>