Use HTTP timeout for redfish:
In the gofish library all calls have to use the
same context that is only allowed to be defined
once with gofish.ConnectContext(). This means that
subsequent calls are bound to that context,
without the ability to modify it (the context field
in the APIClient is not exported and does not have a method to modify it).
This causes issues with "closing" a connection when a context
has timed out or been canceled. This call in redfish
is a call to the server to remove a session. This fixes orphaned
sessions that could lead to the max connections being
reached in the BMC and preventing future calls before the
timeout for these sessions is reached.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>