Update redfish provider to work with HP DL160 redfish:
Setting the boot device on HP DL160 machines was failing.
There were too many values in the payload that were being
rejected. Only sending the BootSourceOverrideEnabled and
BootSourceOverrideTarget got this call working on the HP.

This was working fine before this change against a Dell R240.
With this change the Dell R240 was still working. Hopefully
this is will enable support for more systems generally.

The efiboot setting was also causing the set boot device
redfish call to fail on the HP machine. I didn't seem to
have any affect on the Dell R240 machine so i removed it
entirely.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>