bmc/connection: exclude incompatible provider errors with option
This adds the error type ErrIncompatibleProvider, which may be returned
by Open() in cases where a provider was found to be incompatible with
the device it is connecting to. This error is specifically excluded
from the errors returned by Open() because it adds no real value
and is just noise when seen by the caller, although since it could be useful
for debugging purposes, the option to include the error is given.