Add metadata field for error messages:
This adds the error message for failed method
calls per providers to metadata. These erorrs
were previously not accessible. These errors
are captured in metadata regardless of whether
the method returns successful or not. This is
helpful in debugging providers that error but
the method call succeeds.

The example metadata below shows that while the open method
call succeeded for ipmitool, dell, and gofish it did
not succeed for IntelAMT and asrockrack. We now have
access to the error messages for those failed one.

```
{
    SuccessfulProvider:""
    ProvidersAttempted:[ipmitool asrockrack gofish IntelAMT dell]
    SuccessfulOpenConns:[ipmitool dell gofish]
    SuccessfulCloseConns:[]
    FailedConnDetail:
        map[IntelAMT:provider: IntelAMT: unable to perform digest auth with http://192.168.2.205:16992/wsman: Post "http://192.168.2.205:16992/wsman": dial tcp 192.168.2.205:16992: connect: connection refused
            asrockrack:provider: asrockrack: error unmarshalling response payload: invalid character '<' looking for beginning of value
        ]
}
```

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>