WIP: Add RPC provider:
The idea of the RPC provider is to enable
users the ability to interoperate with existing
bespoke out of band system or as a stop gap for
BMCs/protocols that bmclib does not have a provider
for yet. Internal bespoke out of band systems will
generally not be accepted into BMCLIB.

So users can run a small web endpoint that
implements the RPC contract. The RPC provider sends
a request to the user's web endpoint. The web endpoint
will then translate the call to the out of band
protocol/tool/etc that is needed to communicate with the BMC.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>