Check response in Open method:
Check status code and response payload
in the Open method to validate the endpoint
is conformant with the response contract.

Update checking of the response error code as
well as if it is nil. This will make sure we
dont error out when a response contains a
value for the error instead of just nil.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>