Move back to Go 1.18 in go.mod:
We don't seem to have an offical
strategy for the Go version in
go.mod. I believe that, as a library,
we should only bump this version if
there are dependencies we use that
would require us to move to this version.
I don't believe that we have any dependencies
that warrant a bump.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>