Update virtual media mounting:
At least one BMC, Supermicro SYS-E300-D9, did not
support setting inserted and/or writeProtected
properties in redfish calls to do a virtual media
mount. This falls back to not using them if the
initial call with them in the properties fails.

This was test and worked successfully on a
Supermicro (SYS-E300-D9), HP ILO5, and Dell iDRAC9.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>