expand response payload processing
In contrast to other server-vendors, SMC does not return the task id in
the Location header of the response to a firmware upload. In BMC version
1.05.03 (Redfish version 1.14.0) the payload format changes from a
TaskAccepted message to a Redfish task, which breaks task id detection.
This change adds an attempt to deserialize the task structure before
falling back to the earlier TaskAccepted message type.

This also corrects the startUpdateURI.