Remove sshclient package:
This package isn't used anywhere.
It's in the internal pkg so there isn't
any concerns with breaking clients who
might be using it. Also, there is a high
severity vulnerability in golang.org/x/crypto
that would require us to upgrade to
v0.35.0. This version requires go >= 1.23.0.
We can avoid this by removing sshclient pkg
entirely.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>