Update golangci-lint:
Enable linters in golangci-lint config.
This makes it so that linters are actually
run.

Update golangci-lint version to latest.
Fix all issues after upgrade.

Pass through exit code/status from golangci-lint.
This will cause the Make target to fail with
a non 0 exit. This is needed so that linting is
enforced by CI.

Remove golangci-lint from GitHub Actions.
make all-checks handles downloading and running
golangci-lint and will keep it the version up-to-date.

Signed-off-by: Jacob Weinstock <jakobweinstock@gmail.com>